+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Smartmeter changelog
 
+## v0.3.3 - 2022-01-xx
+* TimescaleDB: replaced \_time\_bucket() with timescaledb\_experimental.time\_bucket\_ng()
+* created yearly\_view
+
 ## v0.3.2 - 2022-01-02
 * set debug level in config file
 * node-red use smartmeter/live topic